Update_Task_14_advanced

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,10 +10,10 @@ int main(void)
 {
 int i;
 int n = 50;
-long long n1 = 0, n2 = 1;
-long long fib;
+long int n1 = 0, n2 = 1;
+long int fib;
 
-printf("%lld, %lld", n1, n2);
+printf("%li, %li", n1, n2);
 
 for (i = 3; i <= n; i++)
 {
@@ -22,7 +22,7 @@ if (i != 51)
 printf(", ");
 }
 fib = n1 + n2;
-printf("%lld", fib);
+printf("%li", fib);
 
 n1 = n2;
 n2 = fib;
